add redis function

+++ sportsCircle/src/main/java/com/cdut/sx/pojo/PageBean.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class PageBean<T> {
     private int currPage;//当前页数
     private int pageSize;//每页显示的记录数
-    private int totalCount;//总记录数
+    private long totalCount;//总记录数
     private int totalPage;//总页数
     private List<T> list;//每页显示的数据
 
@@ -25,11 +25,11 @@ public class PageBean<T> {
         this.pageSize = pageSize;
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public void setTotalCount(int totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
     }
